requirements.txt
asgiref==3.2.7
bcrypt==3.1.7
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
Django==3.0.6
django-cors-headers==3.2.1
idna==2.9
mysqlclient==1.4.6
pycparser==2.20
PyJWT==1.7.1
pytz==2020.1
requests==2.23.0
six==1.14.0
sqlparse==0.3.1
urllib3==1.25.9
